Some bug fixes and some stdlib work

+++ src/test/java/DeleteCompiled.java
@@ -31,12 +31,14 @@ public class DeleteCompiled {
     }
 
     public static void deleteDirectory(File dir) throws IOException {
-        if (dir.isDirectory()) {
-            for (File f : dir.listFiles()) {
-                deleteDirectory(f);
+        if (dir.exists()) {
+            if (dir.isDirectory()) {
+                for (File f : dir.listFiles()) {
+                    deleteDirectory(f);
+                }
             }
+            Files.delete(dir.toPath());
         }
-        Files.delete(dir.toPath());
     }
 
 }
